backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
psycopg2-binary==2.9.1
Flask-CORS==3.0.10 